// ui/src/app/@api-module/model/passwordUpdateHolder.ts
/**
 * Purchases API
 * Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
 *
 * The version of the OpenAPI document: 0.1.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PasswordUpdateHolder { 
    oldPassword?: string;
    newPassword?: string;
}